refactor: Move members folder under management
