+++ app/(club)/club/management/members/components/invitationFormContent.tsx
@@ -3,8 +3,8 @@
 import React, { useState } from "react";
 import Image from "next/image";
 import close from "@/images/icon/close.svg";
-import Alert from "../../../../components/alert/alert";
-import LargeBtn from "../../../../components/button/basicBtn/largeBtn";
+import Alert from "../../../../../components/alert/alert";
+import LargeBtn from "../../../../../components/button/basicBtn/largeBtn";
 import SmallBtn from "@/components/button/basicBtn/smallBtn";
 
 import checkIcon from "@/images/icon/radio_button_checked.svg";
